* 'lsfd-fix-wrong-fmt-str-and-cast-type' of https://github.com/masatake/util-linux:
  lsfd: fix the way to print length of mmap area
  lsfd: fix the way to print stat.st_nlink
  lsfd: fix the way to print stat.st_size

+++ misc-utils/lsfd-file.c
@@ -289,10 +289,10 @@ static bool file_fill_column(struct proc *proc,
 		xasprintf(&str, "%d", (int)file->stat.st_uid);
 		break;
 	case COL_SIZE:
-		xasprintf(&str, "%ju", file->stat.st_size);
+		xasprintf(&str, "%jd", (intmax_t)file->stat.st_size);
 		break;
 	case COL_NLINK:
-		xasprintf(&str, "%ju", (unsigned long int)file->stat.st_nlink);
+		xasprintf(&str, "%ju", (uintmax_t)file->stat.st_nlink);
 		break;
 	case COL_DELETED:
 		xasprintf(&str, "%d", file->stat.st_nlink == 0);
@@ -337,7 +337,7 @@ static bool file_fill_column(struct proc *proc,
 		if (file->association != -ASSOC_SHM
 		    && file->association != -ASSOC_MEM)
 			return true;
-		xasprintf(&str, "%lu", get_map_length(file));
+		xasprintf(&str, "%ju", (uintmax_t)get_map_length(file));
 		break;
 	default:
 		return false;
